feature / Create resource: followed peer advise to show a visual in case user is not authenticated. For that, I isolated the login modal into a small reusable component that I could import in the create resource page
