+++ frontend/src/components/Layout/HeaderComponent.tsx
@@ -1,22 +1,21 @@
-import { useNavigate, useLocation } from "react-router";
+import { useEffect, useRef, useState } from "react";
+import { useLocation, useNavigate } from "react-router";
 import addIcon from "../../assets/add.svg";
-import userIcon from "../../assets/user2.svg";
 import arrowDown from "../../assets/arrow-down.svg";
 import logOutIcon from "../../assets/logOut.svg";
-import ButtonComponent from "../atoms/ButtonComponent";
-import DropdownButtonComponent from "../atoms/DropdownButtonComponent";
+import userIcon from "../../assets/user2.svg";
 import { useUserContext } from "../../context/UserContext";
 import { useChangeUserRole } from "../../hooks/useChangeUserRole";
-import { useEffect, useRef, useState } from "react";
-import { Modal } from "../Modal/Modal";
-import GitHubLogin from "../github-login/GitHubLogin";
-import { AddUsersModal } from "../resources/AddUserModal";
-import { TermsAndConditionsModal } from "../Modal/TermsAndConditionsModal";
-import RoleDropdownComponent from "./header/RoleDropdownComponent";
 import { TypUserRole } from "../../types";
+import { Modal } from "../Modal/Modal";
+import ButtonComponent from "../atoms/ButtonComponent";
+import DropdownButtonComponent from "../atoms/DropdownButtonComponent";
+import { AddUsersModal } from "../resources/AddUserModal";
+import RoleDropdownComponent from "./header/RoleDropdownComponent";
+import { SignInComponent } from "./header/SignInComponent";
 
 const HeaderComponent = () => {
-  const { user, signIn, signOut } = useUserContext();
+  const { user, signOut } = useUserContext();
 
   console.log(user);
   const { isChanging, updateUserRole } = useChangeUserRole();
@@ -25,9 +24,6 @@ const HeaderComponent = () => {
 
   const [resource, setResource] = useState("");
   const [isModalOpen, setIsModalOpen] = useState(false);
-  const [isChecked, setIsChecked] = useState(false);
-  const [loginError, setLoginError] = useState(false);
-  const [isLoading, setIsLoading] = useState(false);
   const [showDropdown, setShowDropdown] = useState(false);
   const [showChangeRoleDropdown, setShowChangeRoleDropdown] =
     useState<boolean>(false);
@@ -40,7 +36,6 @@ const HeaderComponent = () => {
   const langDropdownRef = useRef<HTMLDivElement>(null);
   const roleDropdownRef = useRef<HTMLDivElement>(null);
   const [isAddUserModalOpen, setIsAddUserModalOpen] = useState(false);
-  const [isTermsModalOpen, setIsTermsModalOpen] = useState<boolean>(false);
 
   useEffect(() => {
     const resourcePath =
@@ -99,39 +94,15 @@ const HeaderComponent = () => {
   };
 
   const openModal = () => setIsModalOpen(true);
-  const closeModal = () => setIsModalOpen(false);
 
   const openAddUserModal = () => setIsAddUserModalOpen(true);
   const closeAddUserModal = () => setIsAddUserModalOpen(false);
-
-  const openTermsModal = () => setIsTermsModalOpen(true);
-  const closeTermsModal = () => setIsTermsModalOpen(false);
 
   const userRole: TypUserRole | null = (user?.role as TypUserRole) ?? null;
 
   const hasPermission = userRole
     ? ["superadmin", "admin", "mentor"].includes(userRole)
     : false;
-
-  const handleSignIn = async () => {
-    if (!isChecked) {
-      setLoginError(true);
-      return;
-    }
-    setIsLoading(true);
-    try {
-      await signIn();
-      setIsModalOpen(false);
-    } catch {
-      setLoginError(true);
-    }
-    setIsLoading(false);
-  };
-
-  const handleCheckboxChange = () => {
-    setIsChecked(!isChecked);
-    setLoginError(false);
-  };
 
   return (
     <header className="hidden lg:flex py-4 px-6 items-center justify-end">
@@ -265,63 +236,7 @@ const HeaderComponent = () => {
         )}
 
         {/* MODAL LOGIN */}
-        {isModalOpen && (
-          <Modal closeModal={closeModal} title="Inici de sessió">
-            <GitHubLogin onClick={handleSignIn} isLoading={isLoading} />
-            <section className="flex items-center gap-2 mt-8 font-medium">
-              <label htmlFor="terms">
-                <input
-                  name="terms"
-                  id="terms"
-                  type="checkbox"
-                  onChange={handleCheckboxChange}
-                  checked={isChecked}
-                  className="hidden"
-                />
-                <div
-                  className={`w-5 h-5 flex items-center justify-center rounded border ${
-                    isChecked
-                      ? "bg-[#B91879] border-[#B91879]"
-                      : "border-gray-400"
-                  }`}
-                >
-                  {isChecked && (
-                    <svg
-                      className="w-4 h-4 text-white"
-                      fill="none"
-                      stroke="currentColor"
-                      strokeWidth="2"
-                      viewBox="0 0 24 24"
-                      xmlns="http://www.w3.org/2000/svg"
-                    >
-                      <path
-                        strokeLinecap="round"
-                        strokeLinejoin="round"
-                        d="M5 13l4 4L19 7"
-                      ></path>
-                    </svg>
-                  )}
-                </div>
-              </label>
-              <p>
-                Accepto{" "}
-                <span
-                  className="underline cursor-pointer"
-                  onClick={() => openTermsModal()}
-                >
-                  els termes legals
-                </span>
-              </p>
-            </section>
-
-            {loginError && (
-              <div className="text-red-500 text-[1rem] mt-8 text-center font-medium">
-                Ho sentim, no s'ha pogut iniciar sessió,
-                <br /> contacti amb l'administrador
-              </div>
-            )}
-          </Modal>
-        )}
+        {isModalOpen && <SignInComponent setIsModalOpen={setIsModalOpen} />}
 
         {/* MODAL LOGOUT CONFIRM */}
         {showConfirmLogout && (
@@ -351,14 +266,6 @@ const HeaderComponent = () => {
               </button>
             </div>
           </Modal>
-        )}
-
-        {/* MODAL TERMS AND CONDITIONS */}
-        {isTermsModalOpen && (
-          <TermsAndConditionsModal
-            closeModal={closeTermsModal}
-            title="Termes Legals"
-          />
         )}
 
         {/* MODAL ADD USER */}
